- Simplify STRCAT
- Use CHAR= to compare characters instead of EQ