@sabracrolleton 2018-07-11
Agree with the character comparison. 

On the simplification of strcat, I agree it is simpler, but it does cons more and burn more processor cycles.