@PuercoPop 2018-07-11
Just to check I run an small benchmark on SBCL and you are right. Consing the string upfront has the best performance. Removing said commit.

```lisp
(in-package #:s-sql)

(defparameter +sample-strings+
  (with-open-file (in #P"/home/puercopop/cuis-trie-data.txt")
    (loop :for line := (read-line in nil)
          :while line
          :collect line)))

;; Original Implementation
(defun strcat (args)
  "Concatenate a list of strings into a single one."
  (let ((result (make-string (reduce #'+ args :initial-value 0 :key 'length))))
    (loop :for pos = 0 :then (+ pos (length arg))
          :for arg :in args
          :do (replace result arg :start1 pos))
    result))

(time
 (progn
   (dotimes (i 5000)
     (strcat +sample-strings+))
   (values)))

;; Evaluation took:
;;   10.709 seconds of real time
;;   10.720000 seconds of total run time (10.652000 user, 0.068000 system)
;;   [ Run times consist of 0.268 seconds GC time, and 10.452 seconds non-GC time. ]
;;   100.10% CPU
;;   26,711,897,287 processor cycles
;;   3,041,280,000 bytes consed


;; Using FORMAT
(defun strcat (args)
  "Concatenate a list of strings into a single one."
  (format nil "~{~A~}" args))

(time
 (progn
   (dotimes (i 5000)
     (strcat +sample-strings+))
   (values)))

;; Evaluation took:
;;   15.198 seconds of real time
;;   15.216000 seconds of total run time (15.032000 user, 0.184000 system)
;;   [ Run times consist of 0.340 seconds GC time, and 14.876 seconds non-GC time. ]
;;   100.12% CPU
;;   37,909,358,273 processor cycles
;;   5,572,240,976 bytes consed


;; Using WITH-OUTPUT-TO-STRING
(defun strcat (args)
  "Concatenate a list of strings into a single one."
  (with-output-to-string (out)
    (loop :for arg :in args
          :do (princ arg out))))

(time
 (progn
   (dotimes (i 5000)
     (strcat +sample-strings+))
   (values)))

;; Evaluation took:
;;   12.418 seconds of real time
;;   12.448000 seconds of total run time (12.168000 user, 0.280000 system)
;;   [ Run times consist of 0.488 seconds GC time, and 11.960 seconds non-GC time. ]
;;   100.24% CPU
;;   30,974,924,345 processor cycles
;;   7,856,072,640 bytes consed
```

Where `cuis-trie-data.text` is a ~9000 line text file I had lying around
<details><pre>
&
*
*=
+
+=
,
-
-=
->
/
//
/=
<
<<
<=
=
==
>
>=
>>
@
\\
\\=
\\\
|
~=
~~
a
a11
a11:
a12
a12:
a13
a13:
a21
a21:
a22
a22:
a23
a23:
aaFontsColormapDepth
abandon
abandonChangeText
abandonSources
abbreviation
abbreviation:
aboutThisSystem
aboutToBeGrabbedBy:
aboutToCollapse:
aboutToGrab:
aboutToReturn:through:
abs
abs:arg:
absByteEncode:base:
absolutePathComponentsFor:
absPrintExactlyOn:base:
absPrintOn:base:
absPrintOn:base:digitCount:
absSecure
abstractPC
abstractPCForConcretePC:
abstractSourceMap
abstractSymbolic
accentedLetters
accept:
acceptAndDebugTest
acceptAndDebugTest:
acceptAndTest
acceptAndTest:
acceptAndTestAll
acceptAndTestAll:
acceptAndWithMethodDo:
acceptButton
acceptClicked
acceptContents
acceptContentsFrom:
acceptDroppingMorph:event:
acceptDroppingObject:
acceptedContents
acceptedContentsChanged
acceptedStringOrText
acceptIcon
acceptOnCR
acceptOnCR:
acceptsLoggingOfCompilation
acceptThenTestMethodAndSuite:
accessProtect
actAsExecutor
action
action:
actionForEvent:
actionMap
actionMapFor:
actionMaps
actionsDo:
actionSelector
actionSequenceForEvent:
activate
activate:
activateAndForceLabelToShow
activateAndSendTopToBack:
activateCursor
activateNextWindow
activateOwnerMenu:
activatePreviousWindow
activateReturn:value:
activateSubmenu:
activateWindow
activateWindowAndSendTopToBack:
activeHand
activeHome
activeOuterContext
activePriority
activeProcess
activeSubmenu:
actOnClickFor:
actOnClickFor:in:
actOnClickFor:in:at:
actOnClickFor:in:at:editor:
actualClass
actualContents
actualContents:
actualReceiverClass
actualScope
actualScreenDepth
actualScreenSize
actWhen:
adaptToCollection:andSend:
adaptToComplex:andSend:
adaptToFloat:andSend:
adaptToFraction:andSend:
adaptToInteger:andSend:
adaptToNumber:andSend:
adaptToPoint:andSend:
add:
add:action:
add:action:balloonText:
add:action:icon:
add:action:icon:enabled:
add:after:
add:afterIndex:
add:before:
add:beforeIndex:
add:executor:
add:subMenu:
add:target:action:
add:target:action:argument:
add:target:action:argumentList:
add:target:action:icon:
add:toList:after:
add:withOccurrences:
addAddHandMenuItemsForHalo:hand:
addAdjusterAndMorph:fixedHeight:
addAdjusterAndMorph:layoutSpec:
addAdjusterAndMorph:proportionalHeight:
addAdjusterAndMorph:proportionalWidth:
addAdjusterMorph
addAlarm:after:
addAlarm:with:after:
addAlarm:withArguments:after:
addAlarm:withArguments:for:at:
addAll:
addAllFirst:
addAllLast:
addAllMorphs:
addAllMorphs:after:
addAndClassifySelector:withMethod:inProtocol:notifying:
addArgument:
addAttribute:
addAttribute:from:to:
addAttribute:toArray:
addAttributesForPasting:
addBlankIconsIfNecessary
addButton:to:color:
addButtonFor:
addButtonsTo:color:
addCategory
addCategory:
addCategory:before:
addChildrenForList:addingTo:withExpandedItems:
addClassVarName:
addCoherency:
addCollapseHandle:
addColorMenuItems:hand:
addComment
addContentsTogglesTo:
addCopyItemsTo:
addCustomHaloMenuItems:hand:
addCustomMenuItems:hand:
addDebuggingItemsTo:hand:
addDebugHandle:
addDeferredUIMessage:
addDeltasFrom:
addDependent:
addDependentToHierachy:
addDismissHandle:
addDragHandle:
addDupHandle:
addedChangeRecord
addedChangeRecord:
addedMorph:
addEmbeddingMenuItemsTo:hand:
addEmptyTrailingLine:
addEntry
addEntry:
addEntry:logToFile:
addExportMenuItems:hand:
addExtraMenu2ItemsTo:
addFirst:
addFontEmphHandle:
addFontSizeHandle:
addGetterCodeOn:
addGrabHandle:
addGrowHandle:
addHalo
addHalo:
addHalo:from:
addHaloActionsTo:
addHandle:
addHandles
addHandleSelector
addHandlesTo:box:
addHelpHandle:
addHoistedTemps:
addInstanceVarNamed:withValue:
addInstVarName:
addItem:text:
addItemsFromDictionaries:
addKnownFailing:
addLast:
addLine
addLocalsToPool:
addMenuHandle:
addMethodChange:
addMethodReference:
addMorph:
addMorph:behind:
addMorph:centeredNear:
addMorph:fixedHeight:
addMorph:fixedWidth:
addMorph:inFrontOf:
addMorph:layoutSpec:
addMorph:position:
addMorph:proportionalHeight:
addMorph:proportionalWidth:
addMorphBack:
addMorphBack:position:
addMorphFront:
addMorphFront:position:
addMorphFrontFromWorldPosition:
addMorphKeepMorphHeight:
addMorphs:
addMorphs:widthProportionalTo:
addMorphsTo:from:allowSorting:withExpandedItems:atLevel:
addMorphUseAll:
addNameString:
addNew:byEvaluating:
addNewDirectory
addNewFile
addObsoleteSubclass:
addOptionalHandlesTo:box:
addPossiblyUncoveredAreasIn:to:
addPragma:
addPreamble
addPreference:categories:default:balloonHelp:
addPreference:categories:default:balloonHelp:changeInformee:changeSelector:
addPreference:category:default:balloonHelp:
addPriorVersionsCountForSelector:ofClass:to:
addReadWithin:at:
addRecolorHandle:
addRemoteTemp:encoder:
addRemoteTemp:rootNode:
addRequirement
addressBookIcon
addRotateHandle:
addSelectionRectsFrom:to:to:
addSelector:method:allImplemented:
addSelector:withMethod:
addSelector:withMethod:notifying:
addSelectorSilently:withMethod:
addSelectorTo:
addServiceFor:toMenu:
addServices:for:extraLines:
addSetterCodeOn:
addSharedPool:
addSpecialReference:
addStandardHaloMenuItemsTo:hand:
addStayUpIcons
addString:
addSubclass:
addSubmorphsAfter:fromCollection:allowSorting:
addSystemCategory
addTallies:
addTempNode:
addTest:
addTests:
addTestsFor:toSuite:
addTitle:
addTitleForHaloMenu:
addTo:referencesTo:special:byte:
addToggleItemsToHaloMenu:
addToScroller:
addToShutDownList:
addToShutDownList:after:
addToStartUpList:
addToStartUpList:after:
addToSuite:fromMethods:
addToSuiteFromSelectors:
addUnicodeCodePoint:to:hex:
addUpdating:action:
addUpdating:target:action:
addUpdating:target:action:argumentList:
addWarning:
addWeakDependent:
addWindow:
addWorldHaloMenuItemsTo:hand:
addWriteWithin:at:
adhereTo:
adjustAlarmTimes:
adjustBrightness:
adjustBy:at:
adjustExtent
adjustHorizontallyBy:at:
adjustOwnerAt:
adjustResumptionTimeOldBase:newBase:
adjustSaturation:brightness:
adjustSubmorphsLayout
adjustVerticallyBy:at:
adoptInstance:
adoptInstance:from:
adoptWidgetsColor:
advance
after:
after:ifNone:
afterMidnight:
afterSelectionInsertAndSelect:
alarmScheduler
alarmSortBlock
aleph
align
align:
aligned:with:
alignedTo:
alignment
alignment:
alignmentAt:
alike:
allBehaviorsDo:
allBold
allButFirst
allButFirst:
allButFirstDo:
allButLast
allButLast:
allButLastDo:
allCallsOn
allCallsOn:
allCallsOn:and:
allCategory
allChangeSets
allChangeSetsWithClass:selector:
allChangeTypes
allCharacters
allChildrenDo:
allClasses
allClassesDo:
allClassesImplementing:
allClassVarNames
allContributors
allDirectoriesDo:
allFilesDo:
allFilesDo:matches:
allGlobalRefs
allGlobalRefsWithout:
allImplementedMessages
allImplementedMessagesWithout:
allImplementorsOf:
allImplementorsOf:localTo:
allInstances
allInstancesDo:
allInstancesOrNil
allInstVarNames
allInstVarNamesEverywhere
allLiterals
allLocalCallsOn:
allMask:
allMethodCategoriesIntegratedThrough:
allMethodReferences
allMethodSelectors
allMethodsInCategory:
allMethodsSourceStringMatching:
allMethodsWithSourceString:matchCase:
allMethodsWithString:
allMorphsDo:
allNamed:from:to:
allNamed:from:to:sortedByArgument:
allNamed:from:to:sortedUsing:
allNamed:in:
allNamed:in:sortedByArgument:
allNamed:in:sortedUsing:
allNonWindowRelatedSubmorphs
allObjects
allObjectsDo:
allObjectsOrNil
allocateLiteral:
allowBlockArgumentAssignment
allOwnersDo:
allOwnersReverseDo:
allowsMorphDrop
allowsSubInstVars
allowsSubmorphDrag
allPrimitiveMessages
allPrimitiveMethodsInCategories:
allRegularInstVarNames
allSatisfy:
allSelect:
allSelectors
allSentMessages
allSentMessagesWithout:
allSharedPools
allSubclasses
allSubclassesDo:
allSubclassesDoGently:
allSubclassesWithLevelDo:startingLevel:
allSubclassesWithLevelDo:startingLevel:sortByCategory:
allSubInstances
allSubInstancesDo:
allSuperclasses
allSuperclassesDo:
allSuperclassesUpTo:
allSymbols
allSymbolsIn:do:
allSymbolTablesDo:
allSymbolTablesDo:after:
allTestCaseClassesReferencing:
allTestSelectors
allTestsSending:
allTypicalFileExtensions
allUnimplementedCalls
allUnreferencedClassVariables
allUnreferencedInstanceVariables
allUnSentMessages
allUnsentMessages
allUnSentMessagesIn:
allUnSentMessagesWithout:
allUnusedClassesWithout:
alpha
alpha:
alphabeticalMorphMenu
alphabetizeMessageCategories
alphabetizeSystemCategories
alphaBlendDemo
alphaMask
alphaMask:
alphaMixed:with:
alphaShift
alphaShift:
alter:formBlock:
alternativeBrowseIt
alternativeSourcesName
amountToTranslateWithin:
analogousCodeTo:
analyseArguments:temporaries:rootNode:
analyseClosure:
analyseTempsWithin:rootNode:assignmentPools:
analyze:
analyzedClassName
analyzedClassName:
analyzedClassName:codeSize:instanceCount:spaceForInstances:
anchoredFormOrMorph
anchoredFormOrMorph:
anchorMorph:at:
and
and:
and:and:
and:and:and:
and:and:and:and:
animatedUI
animatedUI:
annotation
annotationForClassCommentFor:
annotationForClassDefinitionFor:
annotationForHierarchyFor:
annotationForSelector:ofClass:
annotationInfo
annotationRequests
annotationSeparator
ansiAssignment
ansiAssignmentPreferenceChanged
anyActive
anyBitOfMagnitudeFrom:to:
anyButtonPressed
anyMask:
anyModifierKey
anyModifierKeyPressed
anyMouseButton
anyOne
anyProcessesAbove:
anySatisfy:
aOrAnPrefix
appearanceIcon
append:
appendContents:
appendExtensionToContentSpec:
appendStream
appendStreamDo:
appendToNew:
appendToString:
appendToText:
appGlobalName
appGlobalName:appSelector:
applyAttribute:
applyConversionMethodsTo:className:varMap:
applySimpleGamma:to:
approxWidthOfText:
appSelector
arcCos
arCosh
arcSin
arcTan
arcTan:
area
areasOutside:
areasOutside:do:
arg
argAdvance:
argument
argumentAt:
argumentCount
argumentGetter:
argumentName
argumentNameAt:havingNamed:
argumentNameSufix
argumentProvider:
arguments
arguments:
arguments:statements:returns:from:
argumentsInEvaluationOrder
argumentTypes
array
arrayIndexForSelection
arrowDown
arrowKey:
arrowLeft
arrowOfDirection:size:
arrowRight
arrowUp
arSinh
arTanh
as8BitColorForm
as:
asAbsolutePathName
asActionSequence
asArray
asBag
asBigCursor
asByteArray
asCamelCase
ascent
asCharacter
asCharacterSet
ascii
asciiValue
asciiValue:
asColorArray
asColorOfDepth:
asComplex
asContentPack:
asContext
asContextWithSender:
asCursorForm
asDate
asDelay
asDictionary
asDigitsAt:in:do:
asDigitsToPower:do:
asDirectoryEntry
asDriveName
asDurationSinceMidnight
asFileEntry
asFileName
asFloat
asFloat64Array
asFloatArray
asFloatPoint
asForm
asFormOfDepth:
asFraction
asGrayForm
asGrayForm:
asGrayScale
asGrayScaleAndTransparent
asGrayScaleAndTransparent:
asHex
asHexDigit
asIdentifier:
asIdentitySet
asIEEE32BitPrecisionFloat
asIEEE32BitWord
asInteger
asIntegerArray
asIntegerOrFloat
asIntegerPoint
askAddedInstVars:
askBeforeDiscardingEdits:
askForCategoryIn:default:
askForDoits
askForSuperclassOf:toImplement:ifCancel:
askForSuperclassOf:upTo:toImplement:ifCancel:
askRemovedInstVars:
askRenames:addTo:using:
asLocal
asLowercase
asMessageSend
asMinimalRepresentation
asMonth
asMouseEnter
asMouseLeave
asMouseMove:
asMouseOver
asNewArray
asNontranslucentColor
asNumber
asOrderedCollection
asPathComponents
asPathTokens
asPoint
asReturnNode
assert
assert:
assert:description:
assert:description:resumable:
assert:equals:
assertIsNotWatching
assertIsWatching
asSet
assignment
assignment:
assignmentCheck:at:
assignmentGlyphSelector
asSmallAngleDegrees
asSmallPositiveDegrees
asSmalltalkComment
associationAt:
associationAt:ifAbsent:
associationForClass
associationOrUndeclaredAt:
associations
associationsDo:
asSortedCollection
asSortedCollection:
asStorableNode:
asString
assureExistence
assureExistenceDirNamed:
assurePostscriptExists
assurePreambleExists
assureSelectionsShow
assureStartupStampLogged
asSymbol
asText
asTrueFraction
asUnaccented
asUnHtml
asUppercase
asUTC
asUtf8
asUtf8:
asValidInstanceVariableName
asValidSelector
asWeek
asWordArray
asYear
at:
at:ifAbsent:
at:ifAbsent:characterIndex:
at:ifAbsentOrNil:
at:ifAbsentOrNilPut:
at:ifAbsentPut:
at:ifPresent:
at:ifPresent:ifAbsent:
at:ifPresentAndInMemory:
at:incrementBy:
at:put:
at:put:characterIndex:
at:setRunOffsetAndValue:
atAll:
atAll:put:
atAll:putAll:
atAllPut:
atClass:add:
atClass:includes:
atEnd
atLast:
atLast:ifAbsent:
atLast:put:
atLeastAsLuminentAs:
atMostAsLuminentAs:
atNewIndex:put:
atPin:
atRandom
atRandom:
atSelector:class:
atSelector:ifAbsent:
atSelector:put:
attachMorph:
attemptToMutateError
attributeArrayForColor:emphasis:
attributesAt:
attributesFor:
attributesFrom:to:do:
atWrap:
atWrap:put:
author
authorInitials
authorInitialsPerSe
authorName
autoBind:
autoCompleterClass
autoCompleterClass:
autoCompleterClassFor:
autoCompletionAround:keyStroke:
autoDeselect:
autoExpand
autoExpand:
automaticPlatformSettings
autoSelectString
autoSelectString:
availableFonts
average
awaitMouseUpIn:repeating:ifSucceed:
awaitMouseUpIn:whileMouseDownDo:whileMouseDownInsideDo:ifSucceed:
back
backChunk
background
backgroundColor:
backgroundColorFillsAllBackground
backgroundImage
backgroundImageData:
backgroundProcess
backspace
backspace:
backTo:
backUpTo:
backWord:
balloonHelp
balloonHelp:
balloonHelpDelayTime
balloonHelpEnabled
balloonText
balloonText:
barycentricCoordinatesInTriangle:with:with:into:ifDegenerate:
base64Decoded
base64Encoded
base:case:
baseFieldList
baseFont
baseFont:
baseKern
baseline
baseName
baseNameAndExtensionFor:do:
baseNameFor:
baseSystemFeature
baseSystemNameFor:
basicActualContents:
basicAt:
basicAt:put:
basicBox
basicCmdShortcutsSpec
basicCompile:notifying:trailer:ifFail:
basicDirectoryExists:
basicDisplayString:from:to:at:strikeFont:
basicInitialize
basicInspect
basicInstallPackageStream:
basicMonitorList
basicNew
basicNew:
basicNext
basicPathComponents:
basicRecursiveDelete
basicRemoveSelector:
basicReplaceAttributesFrom:to:with:
basicReplaceFrom:to:with:
basicScanClosure
basicSetRed:green:blue:
basicSize
basicSuspendedContext
beBindingOfType:announcing:
beBlockArg
beColumn
become:
becomeCompact
becomeCompactSimplyAt:
becomeForward:
becomeForward:copyHash:
becomeUncompact
beCurrent
beDisplay
beep
beExpanded
before:
before:ifNone:
beFullyExpanded
beFullyExpandedAsPer:
beginInstance:size:
beginningOfLine:
beginningOfNextParagraph:
beginningOfParagraph:
beginningOfText
beginReference:
beginsWith:
beginsWithPathSeparator
beginsWithWindowsDriveName
beingAssignedToAnalyseTempsWithin:rootNode:assignmentPools:
beingWaitedOn
beingWaitedOn:
beMethodArg
bench
benchFib
benchmark
benchmarkPrimitiveResponseDelay
beReadOnlyBinding
beReadOnlyBindingAnnouncing:
beReadWriteBinding
beReadWriteBindingAnnouncing:
beRow
beSilent:
beSilentDuring:
beSticky
between:and:
between:and:do:
bigFonts
biggerCursors
biggerNormal
binary
binaryAt:
binaryAt:ifAbsent:
binaryAt:put:
binaryContents
binaryContents:
binaryStream:
bindAndJuggle:
bindArg:
bindBlockArg:within:
bindBlockTemp:
bindBlockTemp:within:
binding
bindingNamesDo:
bindingOf:
bindingReadScanBlockFor:using:
bindingsDo:
bindingWriteScanBlockFor:using:
bindTemp:
bindTemp:in:
bindUndeclaredTemp:
bitAnd:
bitAt:
bitAt:put:
bitClear:
bitInvert
bitInvert16
bitInvert32
bitOr:
bitPatternFor:
bitPatternForDepth:
bitPatternForGrayForm
bitPeekerFromForm:
bitPokerToForm:
bitReverse:
bits
bits:
bitShift:
bitShiftMagnitude:
bitsSize
bitXor:
black
blacker
blank
blankCopyOf:scaledBy:
blankIcon
blend
blendAlpha
blinkParen
blinkParenAt:
blinkPrevParen:
block
blockArgNamesDo:
blockAssociationCheck:
blockDepths
blockDepthsStartIndexes
blockers
blockers:
blockExpression
blockExtent
blockExtentsInto:from:to:scanner:numberer:
blockExtentsToTempsMap
blockForCaseTo:
blockLevelFive
blockLevelFour
blockLevelSeven
blockLevelSix
blockNestingCount
blockNestingCountOf:
blockReturnCode
blockReturnTop
blockScopeRefersOnlyOnceToTemp:
blockTempNamesDo:
blockTemps
blockTo:
blue
blueMask
blueMask:
blueShift
blueShift:
bobsTest
body
bold
boldCode
boldItalic
bonk:with:
bootstrap
border:width:
border:width:borderHeight:fillColor:
border:width:fillColor:
border:width:rule:fillColor:
border:widthRectangle:rule:fillColor:
borderColor
borderColor:
borderWidth
borderWidth:
borderWidth:borderHeight:fillColor:
bottom
bottom:
bottomCenter
bottomContext
bottomLeft
bottomLeftCorner:height:gradientTop:gradientBottom:
bottomLeftCorner:height:gradientTop:gradientBottom:borderWidth:
bottomRight
bottomRightCorner:height:gradientTop:gradientBottom:
bottomRightCorner:height:gradientTop:gradientBottom:borderWidth:
bottomVisibleRowForCanvas:
boundingBox
bounds:
boxExtent
braceArray
braceArray:
braceExpression
braceStream:
braceWith:
braceWith:with:
braceWith:with:with:
braceWith:with:with:with:
braceWithNone
breadthFirstCloseOver:for:
breadthFirstCloseOver:into:for:
breadthFirstScanClosure
break
breakDependents
breakpointMethodSourceFor:in:
brevityState
brightness
bringWindowsFullOnscreen
brown
browse
browse:
browseAllAccessesTo:from:
browseAllCallsOn:
browseAllCallsOn:and:
browseAllCallsOn:localTo:
browseAllCallsOnClass:
browseAllClasses
browseAllImplementorsOf:
…